Add Metrics
Ref: https://paper.dropbox.com/doc/Analytics.js-Metrics-SDD-1hAD90lqGS4aZxHAHYPu7

This makes two main changes:
* Add a metrics client for analytics.js to report metrics to
* Integrate metrics client into analytics.js to report key metrics

The metrics client buffers metrics in memory and periodically uploads them to a server. By default, the metrics client does not upload any metrics (though it points to api.segment.io). It also automatically flushes the buffer when an error metric is recorded. Lastly - the client supports sampling so we can control the metrics being reported.

Metrics are invoked for the following operations:
* A metric is recorded for each method invoked in analytics.js.
* A metric is recorded for each attempt to invoke an analytics.js integration.
* A metric is recorded for each failed attempt when invoking an analytics.js integration.